+++ gendarme/rules/Gendarme.Rules.Interoperability/Test/DelegatesPassedToNativeCodeMustIncludeExceptionHandlingTest.cs
@@ -845,7 +845,6 @@ namespace Test.Rules.Interoperability
     }
 
     [Test]
-    [Ignore("False positive w/stack underflow")]
     public void Test_EventOperator()
     {
       AssertTest("HandleEventOperatorOK", 0);