+++ ReleaseNotes.md
@@ -1,5 +1,6 @@
-# 2022.xx.xx.xxxxx
+# 2022.5.15.122xx
 
 * Fix some false negatives for `AvoidSwitchStatementsRule` (where the switch is on a `string` value)
+* Fix false positives for `DelegatesPassedToNativeCodeMustIncludeExceptionHandlingRule` and `AvoidUnneededUnboxingRule` associated with changes in F# exception handling
 
 ⁋For previous releases, go here -- https://github.com/SteveGilham/Gendarme/blob/trunk/ReleaseNotes%20-%20Previously.md